change in the handling of the Display Activation message (as well as the VT Status message).  This is a "proposal", made to revision 3665 of comm/Part6_VirtualTerminal_Client/impl.
The handling of the Display Activation message had a delayed action in
some cases - particularly, when the client's state was already active.
In that case, no change would be recorded - instead, the update of the
vtclientservercommmunication object would wait for the next VT Status
message (with the possible issue that hidden states may be lost when
there are more WS than possible windows).

Another issue - the one the precipitated this change - is that data
available in the Display Activation message would have to be stored for
a future timeframe in order to be delivered in the callback, which would
be thusly delayed until the VT Status message.

This change has only been superficially tested (having been researched
and implemented on Good Friday and the ensuing Saturday).  Thus, it's
only a proposal at this point.


git-svn-id: https://projects.osb-ag.de/svn/OSB/IsoAgLib/IsoAgLib@3666 2a0317ad-70bf-0310-b33b-a4ce7cd1c90a