[Bugfix] [CAN-HALs: Standard Identifier processing] [Standard Identifier CAN-messages weren't handled properly]
In case of STD messages the Msg-ID had to be shifted by 18 bits,
as it was placed into the upper 11 bit of the 29 bit ID field.
Hence the messages were inserted into the central-fifo, but the
FilterBox_c-processing couldn't properly match the IDs.

Done by Achim Spangler & Mike Schmidt
Review: Martin Wodok


git-svn-id: https://projects.osb-ag.de/svn/OSB/IsoAgLib/IsoAgLib@3967 2a0317ad-70bf-0310-b33b-a4ce7cd1c90a