[VT2ISO] add option to provide a xml search path ( -d= | -D= )
Possibility to use some xml files in the  project file mode that are in variable locations (e.g generated xml's during
build process and build dir != src dir ).

The command line switches -d= or -D give a list seperated by ":" with relative or absolute pathes, that are used to
search for xml file marked as relative within a vtp. If no file with the corresponding name is found, the project
directory (thats where the vtp resides ) is prepended to the file name.

Attention: Usage() does not tell about the switch to avoid confusion!

Done by: Felix Obenhuber
Reviewed by: Martin Wodok



git-svn-id: https://projects.osb-ag.de/svn/OSB/IsoAgLib/IsoAgLib@4115 2a0317ad-70bf-0310-b33b-a4ce7cd1c90a