[Enhancement][Cleanup][Updated iIdentItem_c interface][Added EEprom tutorial]
A iIdentitem is no longer initialized via constructor. Parameters can be passed only
with the init call. Furthermore the parameters are grouped by the usage of iIsoname_c.

The storage of a successfully claimed SA is fixed and now handled by a instance of
a iIdentDataStorage_c implementation. The tutorial extended with a eeprom tutorial
and the correct usage of the new iIdentItem_c interface.

Reviewed by: M.Wodok

git-svn-id: https://projects.osb-ag.de/svn/OSB/IsoAgLib/IsoAgLib@4768 2a0317ad-70bf-0310-b33b-a4ce7cd1c90a